refactor(media): split the engine-neutral half out of the Mux source
The Mux source layer is almost entirely about playback identity and the URLs
derived from it — engine-neutral work that every Mux Media needs. But it sat in
one module with `MuxSource extends HlsSource` and a `createMuxDrmSystems` that
returns hls.js's `DRMSystemsConfiguration`, so importing any of it meant
importing hls.js's types. An SPF-backed Mux Media can't do that: `@videojs/spf`
has no hls.js dependency, and shouldn't gain one.

Split along that seam. `mux/source/` holds the neutral layer — `MuxSourceBase`,
the params types, and the stream/poster/storyboard URL builders — behind its own
entry point, `@videojs/media/dom/mux/source`, whose emitted declarations import
nothing. `mux/drm.ts` keeps the hls.js license-server shaping, now taking
`MuxSourceBase` since it only ever read Mux identity. `MuxSource` itself moves
beside `MuxMedia`, the Media whose engine it describes, as
`HlsSource & MuxSourceBase`.

`@videojs/media/dom/mux` re-exports the neutral layer, so the public surface is
unchanged and `MuxMedia` keeps one import site. Behavior-preserving; the
existing Mux tests are the guard, with the DRM cases moving to `tests/drm.test.ts`
alongside what they now cover.
