+++ packages/media/src/dom/mux/index.ts
@@ -1,5 +1,9 @@
 export * from './media';
 export { MuxData, type MuxDataProps, muxDataDefaultProps } from './mux-data';
+// The engine-neutral source layer, re-exported so this stays the one import for
+// the hls.js-backed Media. Its own entry point, `@videojs/media/dom/mux/source`,
+// is what a Media on another engine imports — reaching it through here would
+// pull hls.js in with it.
 export type {
   MuxDrmParams,
   MuxImageExt,
@@ -8,6 +12,6 @@ export type {
   MuxPosterParams,
   MuxRenditionOrder,
   MuxResolution,
-  MuxSource,
+  MuxSourceBase,
   MuxStoryboardParams,
-} from './utils';
+} from './source';
