+++ packages/media/src/dom/mux/source/index.ts
@@ -1,0 +1,1 @@
+export * from './source';
